backend/requirements.txt
flask
flask-cors
apscheduler
requests
python-dotenv
